refactor: extract xtdb query helpers
