+++ polars_hist_db/backends/xtdb_transport.py
@@ -15,11 +15,17 @@ from ..types import PolarsType, is_polars_type
 from .temporal import system_time_hint_clause
 
 
-_XTDB_SYSTEM_COLUMNS = {"_valid_from", "_valid_to", "_system_from", "_system_to"}
 _XTDB_LAST_SYSTEM_TIME_KEY = "polars_hist_db_xtdb_last_system_time"
 _XTDB_RESERVED_IDENTIFIERS = {"flag", "timestamp"}
 _XTDB_QUERY_ROWS_PER_CHUNK = 10_000
 _IDENTIFIER_RE = re.compile(r"^[A-Za-z_][A-Za-z0-9_]*$")
+
+
+def __getattr__(name: str) -> Any:
+    """Keep legacy private query imports working during the module split."""
+    from . import xtdb_query
+
+    return getattr(xtdb_query, name)
 
 
 def _table_config_ops(connection: Any) -> Any:
@@ -326,6 +332,18 @@ def _xtdb_insert_casts(
     return casts
 
 
+def _xtdb_polars_type_or_none(data_type: str) -> pl.DataType | None:
+    normalized = data_type.upper()
+    if normalized.startswith(("DATETIME", "TIMESTAMPTZ")) or normalized.startswith(
+        "TIMESTAMP WITH TIME ZONE"
+    ):
+        return pl.Datetime("us", "UTC")
+    try:
+        return PolarsType.from_sql(data_type)
+    except ValueError:
+        return None
+
+
 def _xtdb_configured_column_dtypes(
     table_config: TableConfig,
 ) -> dict[str, pl.DataType]:
@@ -615,124 +633,6 @@ def _execute_xtdb_arrow_copy(
             driver_connection.autocommit = autocommit
 
 
-def _xtdb_temporal_basis_clause(update_time: Optional[datetime]) -> str:
-    if update_time is None:
-        return ""
-    return (
-        f" FOR VALID_TIME AS OF {_xtdb_timestamp_literal(update_time)}"
-        f" FOR SYSTEM_TIME AS OF {_xtdb_timestamp_literal(update_time)}"
-    )
-
-
-def _xtdb_valid_time_clause(time_hint: Optional[TimeHint]) -> str:
-    if time_hint is None or time_hint.mode == "none":
-        return ""
-    if time_hint.mode == "all":
-        return " FOR VALID_TIME ALL"
-    if time_hint.mode == "asof":
-        assert isinstance(time_hint.asof_utc, datetime)
-        return f" FOR VALID_TIME AS OF {_xtdb_timestamp_literal(time_hint.asof_utc)}"
-    if time_hint.mode == "span":
-        assert isinstance(time_hint.asof_utc, datetime)
-        assert time_hint.history_span is not None
-        if time_hint.history_span.total_seconds() == 0:
-            return (
-                f" FOR VALID_TIME AS OF {_xtdb_timestamp_literal(time_hint.asof_utc)}"
-            )
-        start_date_utc = time_hint.asof_utc - time_hint.history_span
-        return (
-            " FOR VALID_TIME BETWEEN "
-            f"{_xtdb_timestamp_literal(start_date_utc)} AND "
-            f"{_xtdb_timestamp_literal(time_hint.asof_utc)}"
-        )
-
-    raise ValueError(f"invalid TimeHint mode: {time_hint.mode}")
-
-
-def _xtdb_single_primary_key_alias(table_config: TableConfig) -> str | None:
-    primary_keys = list(table_config.primary_keys)
-    if len(primary_keys) != 1:
-        return None
-    primary_key = primary_keys[0]
-    if primary_key != "_id":
-        return primary_key
-    return None
-
-
-def _xtdb_table_query_output_columns(
-    table_config: TableConfig,
-    column_selection: Optional[list[str]],
-) -> list[str]:
-    if column_selection is not None:
-        return column_selection
-
-    columns = []
-    single_key_alias = _xtdb_single_primary_key_alias(table_config)
-    if single_key_alias is not None:
-        columns.append(single_key_alias)
-
-    columns.extend(
-        column.name
-        for column in table_config.columns
-        if column.name not in {"_id", single_key_alias, *_XTDB_SYSTEM_COLUMNS}
-    )
-    columns.extend(["__valid_from", "__valid_to"])
-    return columns
-
-
-def _xtdb_table_query_select_expr(column: str, table_config: TableConfig) -> str:
-    single_key_alias = _xtdb_single_primary_key_alias(table_config)
-    if single_key_alias == column:
-        return f"t._id AS {_xtdb_column_identifier(column)}"
-    if column == "__valid_from":
-        return "t._valid_from AS __valid_from"
-    if column == "__valid_to":
-        return "t._valid_to AS __valid_to"
-    return f"t.{_xtdb_column_identifier(column)}"
-
-
-def _xtdb_table_query_target_column(column: str, table_config: TableConfig) -> str:
-    if _xtdb_single_primary_key_alias(table_config) == column:
-        return "_id"
-    return _xtdb_column_identifier(column)
-
-
-def _xtdb_values_cte(name: str, df: pl.DataFrame) -> tuple[str, dict[str, Any]]:
-    if df.is_empty():
-        raise ValueError("XTDB table_query requires at least one query row")
-
-    cte_name = _validate_identifier(name)
-    columns = [_xtdb_column_identifier(column) for column in df.columns]
-    casts = [_xtdb_cast_type_from_polars(dtype) for dtype in df.schema.values()]
-    parameters = {}
-    row_queries = []
-    for row_index, row in enumerate(df.rows()):
-        projections = []
-        for column_index, (value, cast_type) in enumerate(zip(row, casts, strict=True)):
-            parameter = f"q_{row_index}_{column_index}"
-            parameters[parameter] = _xtdb_parameter_value(value, cast_type)
-            projections.append(
-                f"CAST(:{parameter} AS {cast_type}) AS {columns[column_index]}"
-            )
-        row_queries.append(f"SELECT {', '.join(projections)}")
-    return (
-        f"{cte_name} AS ({' UNION ALL '.join(row_queries)})",
-        parameters,
-    )
-
-
-def _xtdb_polars_type_or_none(data_type: str) -> pl.DataType | None:
-    normalized = data_type.upper()
-    if normalized.startswith(("DATETIME", "TIMESTAMPTZ")) or normalized.startswith(
-        "TIMESTAMP WITH TIME ZONE"
-    ):
-        return pl.Datetime("us", "UTC")
-    try:
-        return PolarsType.from_sql(data_type)
-    except ValueError:
-        return None
-
-
 class XtdbDataframeOps:
     def __init__(
         self,
@@ -797,6 +697,16 @@ class XtdbDataframeOps:
         table_config: TableConfig | None = None,
         basis_time: datetime | None = None,
     ) -> pl.DataFrame:
+        from .xtdb_query import (
+            _xtdb_single_primary_key_alias,
+            _xtdb_table_query_output_columns,
+            _xtdb_table_query_select_expr,
+            _xtdb_table_query_target_column,
+            _xtdb_temporal_basis_clause,
+            _xtdb_valid_time_clause,
+            _xtdb_values_cte,
+        )
+
         if table_config is None:
             table_config = _table_config_ops(self.connection).from_table(
                 table_schema,
